requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.1
attrs==25.3.0
audioop-lts==0.2.2
brotli==1.2.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.122.0
ffmpy==1.0.0
filelock==3.20.0
filetype==1.2.0
firebase-admin==6.8.0
flatbuffers==25.9.23
fonttools==4.59.2
frozenlist==1.6.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gradio==5.44.1
gradio_client==1.12.1
gradio_rangeslider==0.0.8
greenlet==3.2.1
groovy==0.1.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.4
mne==1.10.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
mypy_extensions==1.1.0
numpy==2.3.5
onnx==1.19.1
onnxruntime==1.23.2
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
pooch==1.8.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.10
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pyreadline3==3.5.4
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
ruff==0.14.7
safehttpx==0.1.7
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
zstandard==0.23.0
